Refactor job CLI into package
